Readability improvements
I added more comprehensive doc strings and comments for both the main file and the input handler file. Speaking of the input handler file. I named the file input_handler.py because previously it was InputIO.py, which made no sense. So I've changed the name to make it's purpose clearer for anyone trying to understand the code.